requirements.txt
absl-py==1.2.0
accelerate==0.12.0
aif360==0.5.0
aiohttp==3.7.4.post0
alembic==1.8.1
antlr4-python3-runtime==4.8
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asttokens==2.0.8
async-timeout==3.0.1
attrdict==2.0.1
attrs==22.1.0
autopage==0.5.1
awscli==1.27.8
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.24.51
botocore==1.29.8
bravado==11.0.3
bravado-core==5.17.0
brotlipy==0.7.0
cachetools==5.2.0
captum==0.4.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
cliff==4.0.0
cmaes==0.8.2
cmd2==2.4.2
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.6.0
commonmark==0.9.1
contractions==0.1.73
cryptography @ file:///tmp/build/80754af9/cryptography_1652083738073/work
cycler==0.11.0
dacite==1.6.0
datasets==2.6.1
debugpy==1.6.3
decorator==5.1.1
deepspeed==0.7.2
defusedxml==0.7.1
dill==0.3.5.1
docformatter==1.5.1
docopt==0.6.2
docutils==0.16
dython==0.7.2
emoji==2.2.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.0.0
fairscale==0.4.13
fastjsonschema==2.16.2
filelock==3.8.0
flake8==6.0.0
flake8-bugbear==22.12.6
fonttools==4.34.4
fqdn==1.5.1
fsspec==2022.10.0
future==0.18.2
fuzzywuzzy==0.18.0
gitdb==4.0.9
gitdb2==4.0.2
GitPython==3.1.27
google-api-core==2.11.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-cloud-storage==2.7.0
greenlet==1.1.2
grpcio==1.47.0
hjson==3.1.0
huggingface-hub==0.8.1
hydra-colorlog==1.1.0
hydra-core==1.1.0
hydra-optuna-sweeper==1.1.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.12.0
importlib-resources==5.9.0
iopath==0.1.10
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonlines==1.2.0
jsonpickle==1.5.2
jsonpointer==2.3
jsonref==0.2
jsonschema==4.9.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras-tuner==1.1.3
kiwisolver==1.4.4
krippendorff==0.5.1
kt-legacy==1.0.4
lightning-transformers==0.2.1
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
monotonic==1.6
moto==4.0.9
msgpack==1.0.4
multidict==5.2.0
multiprocess==0.70.12.2
munch==2.5.0
myst-parser==0.12.10
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.6.1
neptune-client==0.14.2
neptune-contrib==0.28.1
nest-asyncio==1.5.6
ninja==1.10.2.3
nltk==3.7
notebook==6.4.12
numpy==1.19.5
oauthlib==3.2.0
omegaconf==2.1.1
openai==0.23.0
openpyxl==3.0.10
optuna==2.4.0
packaging==21.3
pandas==1.4.3
pandas-stubs==1.4.4.220906
pandocfilters==1.5.0
parallelformers==1.2.7
-e git+https://github.com/facebookresearch/ParlAI.git@3f70d6b0979934c7597b34baad3b76eb86b64c74#egg=parlai
parso==0.8.3
patsy==0.5.3
pbr==5.10.0
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
portalocker==2.6.0
prettytable==3.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==8.0.0
py-gfm==2.0.0
py-rouge==1.1
pyaml==21.10.1
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.2
pyDeprecate==0.3.1
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==7.2.0
pytest-regressions==2.4.1
python-dateutil==2.8.2
pytorch-lightning==1.6.0
pytz==2022.2.1
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
rapidfuzz==2.13.7
regex==2022.7.25
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-mock==1.10.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==10.6.0
rsa==4.7.2
s3transfer==0.6.0
sacred==0.8.2
scikit-learn==1.0.2
scikit-optimize==0.9.0
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.12.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
sh==1.14.3
simcse==0.4
simpledorff==0.0.2
simplejson==3.17.6
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
soupsieve==2.3.2.post1
Sphinx==2.2.2
sphinx-autodoc-typehints==1.10.3
sphinx-rtd-theme==1.2.0rc1
SQLAlchemy==1.4.40
stack-data==0.5.0
statsmodels==0.13.2
stevedore==4.0.0
strict-rfc3339==0.7
subword-nmt==0.3.8
swagger-spec-validator==2.7.4
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1+cu102
torchaudio==0.12.1
torchmetrics==0.7.2
torchtext==0.13.1
torchvision==0.13.1+cu102
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.21.1
types-pytz==2022.2.1.0
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
Unidecode==1.3.6
uri-template==1.2.0
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xgboost==1.6.2
xmltodict==0.13.0
xxhash==2.0.2
yarl==1.7.0
yellowbrick==1.5
zipp==3.8.1
